run-requirements.txt
Flask==2.3.3
flask-cors
google-cloud-pubsub==2.21.0
